% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNumberPC.R
\name{FindNumberPC}
\alias{FindNumberPC}
\alias{FindNumberPC.Seurat}
\title{FindNumberPC}
\usage{
FindNumberPC(object = NULL, dim2Test = NULL, returnBIC = FALSE, ...)

\method{FindNumberPC}{Seurat}(object = NULL, dim2Test = NULL, returnBIC = FALSE, ...)
}
\arguments{
\item{object}{object for which to find the number of PCs. (options: Seurat, pca)}

\item{dim2Test}{Number of dimensions to test}

\item{returnBIC}{Return the BIC values used to select the number of PCs.}
}
\value{
The selected number of PCs.
}
\description{
FindNumberPC
}
