% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicPCA.R
\name{getBic}
\alias{getBic}
\title{getBic}
\usage{
getBic(eig = NULL, N = NULL, d = NULL, k = NULL)
}
\arguments{
\item{eig}{Eigenvalues of the covariance matrix}

\item{N}{Number of observations}

\item{d}{Original number of dimensions}

\item{k}{Selected dimensions to calculate the BIC}
}
\value{

}
\description{
Get the the BIC for a defined number of PCs
}
