% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BicPCA.R
\name{getBicPca}
\alias{getBicPca}
\title{getBicPca}
\usage{
getBicPca(pcaData = NULL, eig = NULL, K = NULL, N = NULL)
}
\arguments{
\item{pcaData}{PCA object}

\item{K}{Number of PCs to test}
}
\value{

}
\description{
Get the BIC for PCA
}
